requirements.txt
asgiref==3.6.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.8.0rc2
distlib==0.3.6
Django==4.1.8
django-cors-headers==4.3.1
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
dnspython==2.3.0
et-xmlfile==1.1.0
filelock==3.12.0
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
Pillow==9.5.0
platformdirs==3.2.0
psycopg2-binary
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3
urllib3==2.2.0
virtualenv==20.22.0
whitenoise==6.6.0
pymongo[srv]